Added hystheresis support

+++ VisualStudio/LaneKeeping/main.cpp
@@ -16,16 +16,15 @@ int main(int argc, const char** argv) {
     if (!cap.isOpened()) return -1;
 
     //Set up lane detection parameters
-    cv::Size frameSize = cv::Size(cap.get(cv::CAP_PROP_FRAME_WIDTH), cap.get(cv::CAP_PROP_FRAME_HEIGHT));
+    cv::Size frameSize = cv::Size( static_cast<int>( cap.get(cv::CAP_PROP_FRAME_WIDTH) ), static_cast<int>( cap.get(cv::CAP_PROP_FRAME_HEIGHT) ) );
     double frameFormat = cap.get(cv::CAP_PROP_FORMAT);
     LaneDetection::prepare(frameSize, frameFormat);
 
     cv::Mat frame;
 
-    auto globalTimer = new Timer("Loop", cap.get(cv::CAP_PROP_FRAME_COUNT));
+    auto globalTimer = new Timer("Loop", static_cast<long long>( cap.get(cv::CAP_PROP_FRAME_COUNT) ) );
 
     while (cap.read(frame)) {
-
         auto t = new Timer("Loop");
 
         LaneDetection::process(frame);
